+++ kittens/diff/main.py
@@ -89,6 +89,11 @@ If an invalid expression is provided, diff will fail with an error.
 '''
     )
 
+opt('sticky_header', 'no', option_type='to_bool', long_text='''
+When scrolled past the header of a file, keep the file's name and separator
+pinned to the top of the screen.
+''')
+
 egr()  # }}}
 
 # colors {{{